# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /stations:
    get:
      operationId: getStations
      summary: List stations with optional filters
      parameters:
        - in: query
          name: connector_type
          schema:
            type: string
        - in: query
          name: min_power
          schema:
            type: number
        - in: query
          name: max_power
          schema:
            type: number
        - in: query
          name: status
          schema:
            type: string
            enum: [free, occupied, offline]
        - in: query
          name: operator_id
          schema:
            type: integer
        - in: query
          name: lat
          schema:
            type: number
        - in: query
          name: lng
          schema:
            type: number
        - in: query
          name: radius_km
          schema:
            type: number
        - in: query
          name: amenities
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Station"
    post:
      operationId: createStation
      summary: Create station (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"
  /stations/{id}:
    get:
      operationId: getStation
      summary: Get station by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"
    put:
      operationId: updateStation
      summary: Update station (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StationInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"
    delete:
      operationId: deleteStation
      summary: Delete station (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /stations/{id}/status:
    patch:
      operationId: updateStationStatus
      summary: Toggle station status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StationStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"

  # ── Vehicles ─────────────────────────────────────────────────
  /vehicles:
    get:
      operationId: getVehicles
      summary: List vehicles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      summary: Create a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      summary: Get vehicle by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    put:
      operationId: updateVehicle
      summary: Update vehicle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      summary: Delete vehicle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Routes ───────────────────────────────────────────────────
...
  /users:
    get:
      operationId: getUsers
      summary: List all users (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      summary: Create or register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      summary: Update user profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Payment Methods ───────────────────────────────────────────
...
  /support-tickets:
    get:
      operationId: getSupportTickets
      summary: List support tickets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support-tickets/{id}:
    patch:
      operationId: updateSupportTicket
      summary: Update ticket status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
components:
  schemas:
...
    Operator:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        logo_url:
          type: ["string", "null"]
        station_count:
          type: integer
...
    Connector:
      type: object
      required: [type, power_kw, total, available]
      properties:
        type:
          type: string
          enum: [CCS2, CHAdeMO, Type2, GB-T]
        power_kw:
          type: number
        total:
          type: integer
        available:
          type: integer

    # ── Station ───────────────────────────────────────────────
    Station:
      type: object
      required: [id, name, address, lat, lng, power_kw, price_per_kwh, status, source]
      properties:
        id:
          type: integer
        operator_id:
          type: ["integer", "null"]
        operator:
          $ref: "#/components/schemas/Operator"
        name:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        connectors:
          type: array
          items:
            $ref: "#/components/schemas/Connector"
        power_kw:
          type: number
        price_per_kwh:
          type: number
        status:
          type: string
          enum: [free, occupied, offline]
        source:
          type: string
          enum: [manual, api, mock]
        amenities:
          type: array
          items:
            type: string
        updated_at:
          type: string
          format: date-time
        distance_km:
          type: ["number", "null"]
    StationInput:
      type: object
      required: [name, address, lat, lng, power_kw, price_per_kwh, status]
      properties:
        operator_id:
          type: ["integer", "null"]
        name:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        connectors:
          type: array
          items:
            $ref: "#/components/schemas/Connector"
        power_kw:
          type: number
        price_per_kwh:
          type: number
        status:
          type: string
          enum: [free, occupied, offline]
        source:
          type: string
          enum: [manual, api, mock]
        amenities:
          type: array
          items:
            type: string
    StationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [free, occupied, offline]

    # ── Vehicle ───────────────────────────────────────────────
    Vehicle:
      type: object
      required: [id, name, battery_kwh, range_km, connector_type]
      properties:
        id:
          type: integer
        name:
          type: string
        battery_kwh:
          type: number
        range_km:
          type: number
        connector_type:
          type: string
          enum: [CCS2, CHAdeMO, Type2, GB-T]
        current_battery_pct:
          type: ["number", "null"]
    VehicleInput:
      type: object
      required: [name, battery_kwh, range_km, connector_type]
      properties:
        name:
          type: string
        battery_kwh:
          type: number
        range_km:
          type: number
        connector_type:
          type: string
          enum: [CCS2, CHAdeMO, Type2, GB-T]
        current_battery_pct:
          type: ["number", "null"]

    # ── Route Stop ────────────────────────────────────────────
...
    User:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        membership_tier:
          type: string
          enum: [free, premium]
        total_sessions:
          type: integer
        total_spent:
          type: number
        total_energy_kwh:
          type: number
        co2_saved_kg:
          type: number
        created_at:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        membership_tier:
          type: string
          enum: [free, premium]

    # ── Payment Method ────────────────────────────────────────
...
    SupportTicket:
      type: object
      required: [id, subject, message, status, created_at]
      properties:
        id:
          type: integer
        user_id:
          type: ["string", "null"]
        user_email:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: ["string", "null"]
          format: date-time
    SupportTicketInput:
      type: object
      required: [subject, message]
      properties:
        user_id:
          type: ["string", "null"]
        user_email:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
    SupportTicketUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]