# lib/api-zod/src/generated/types/stationStatusUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { StationStatusUpdateStatus } from './stationStatusUpdateStatus';

export interface StationStatusUpdate {
  status: StationStatusUpdateStatus;
}