# lib/api-zod/src/generated/types/supportTicketUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { SupportTicketUpdateStatus } from './supportTicketUpdateStatus';

export interface SupportTicketUpdate {
  status?: SupportTicketUpdateStatus;
}