# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { UserMembershipTier } from './userMembershipTier';

export interface User {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  phone?: string | null;
  membership_tier?: UserMembershipTier;
  total_sessions?: number;
  total_spent?: number;
  total_energy_kwh?: number;
  co2_saved_kg?: number;
  created_at?: Date;
}